refactor(project): split cards in files
